# swagger/api-spec.json
{"openapi":"3.0.0","paths":{"/":{"get":{"operationId":"AppController_getHello","parameters":[],"responses":{"200":{"description":""}},"tags":["App"]}},"/merchant":{"post":{"operationId":"createMerchant","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMerchantDto"}}}},"responses":{"201":{"description":""}},"summary":"","tags":["Merchant"]}},"/payment/status":{"get":{"operationId":"checkPayerStatus","parameters":[{"name":"address","required":true,"in":"query","schema":{"type":"string"}},{"name":"invoiceId","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"200":{"description":""}},"summary":"","tags":["Payment"]}},"/payment/invoice":{"post":{"operationId":"sendInvoicePayment","parameters":[{"name":"payer","required":true,"in":"query","schema":{"type":"string"}},{"name":"invoiceId","required":true,"in":"query","schema":{"type":"number"}}],"responses":{"201":{"description":""}},"summary":"","tags":["Payment"]}},"/payment/asset":{"post":{"operationId":"convertThenSendStable","parameters":[{"name":"payer","required":true,"in":"query","schema":{"type":"string"}},{"name":"invoiceId","required":true,"in":"query","schema":{"type":"number"}},{"name":"assetAddress","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"summary":"","tags":["Payment"]}},"/invoice":{"post":{"operationId":"createInvoice","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateInvoiceDto"}}}},"responses":{"201":{"description":""}},"summary":"","tags":["Invoice"]},"get":{"operationId":"findAllInvoice","parameters":[],"responses":{"200":{"description":""}},"summary":"","tags":["Invoice"]}},"/invoice/{id}":{"get":{"operationId":"findInvoiceById","parameters":[{"name":"id","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":""}},"summary":"","tags":["Invoice"]}},"/token":{"post":{"operationId":"createToken","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateTokenDto"}}}},"responses":{"201":{"description":""}},"summary":"","tags":["Token"]}},"/token/faucet":{"post":{"operationId":"mintToken","parameters":[{"name":"to","required":true,"in":"query","schema":{"type":"string"}},{"name":"amount","required":true,"in":"query","schema":{"type":"number"}},{"name":"tokenAddress","required":true,"in":"query","schema":{"type":"string"}}],"responses":{"201":{"description":""}},"summary":"","tags":["Token"]}}},"info":{"title":"PayNeu API","description":"","version":"0.1","contact":{}},"tags":[],"servers":[],"components":{"schemas":{"CreateMerchantDto":{"type":"object","properties":{"name":{"type":"string","example":"PayNeu Technology"}},"required":["name"]},"CreateInvoiceDto":{"type":"object","properties":{"details":{"type":"string","example":"Payment for buying a digital product"},"merchant_id":{"type":"number","example":1},"token_id":{"type":"number","example":1},"amount":{"type":"number","example":100}},"required":["details","merchant_id","token_id","amount"]},"CreateTokenDto":{"type":"object","properties":{"address":{"type":"string","example":"0x35435120c2cf51f7f122f2b37bda3bbc686831de"},"name":{"type":"string","example":"Bazed Token"}},"required":["address","name"]}}}}